refactor(lua): plugins as plain Lua modules under lua/plugin/
`runtime/plugin/<name>.lua` had been the last special directory
on the Lua side: bundled `ttymap.plugin_searcher` walked
`<layer>/plugin/<name>.lua` via a custom `package.searchers`
entry, while every other Lua file resolved through standard
`package.path`. After dropping the plugin attribution wrap in
#284, that custom searcher was no longer pulling its weight —
plugins are just `.lua` files that call `register_*`.

Move plugins under `runtime/lua/plugin/<name>.lua` (or
`<name>/init.lua`) so they reach via the same `package.path`
resolution as `ttymap.fmt` / `ttymap.notify`. `runtime/init.lua`
activates each as `require "plugin.<name>"`. Sub-modules
(`require "plugin.travel.routes.italy"`) work via the same
mechanism, no special precedence rules.

Drops:
- `runtime/lua/ttymap/plugin_searcher.lua` (custom searcher lib)
- The bundled `runtime/init.lua`'s
  `require("ttymap.plugin_searcher").install()` call

What remains:
- `ttymap.runtime_path` is still exposed on the Lua side,
  informational; `vm::new_lua` already prepends each layer's
  `lua/` to `package.path`, so plugin / lib resolution Just
  Works for any layer in the runtime path.
- `<layer>/lua/plugin/<name>.lua` vs `<layer>/lua/ttymap/<name>.lua`
  is a pure Lua-side convention: plugins call `register_*` for
  side effects, libs return module tables. The host treats both
  identically.

Migration:
- 16 bundled plugins moved (git renames, no diff).
- Sub-module requires updated: `require "aircraft.opensky"` →
  `require "plugin.aircraft.opensky"`, similar for satellite,
  search, travel.
- Makefile install-runtime: drop `cp runtime/plugin/...`; only
  `lua/` + `init.lua` ship now. Old `~/.local/share/ttymap/plugin/`
  is wiped on first reinstall.
- Test helper `fresh_pluginsearcher_state` →
  `fresh_lua_layer_state`; tests call
  `require "plugin.<name>"` instead of bare `require "<name>"`.

User migration: any `~/.config/ttymap/plugin/` user plugins must
move to `~/.config/ttymap/lua/plugin/`; init.lua activations
become `require "plugin.<name>"`.

Doc updates: README, CONTRIBUTING, CLAUDE.md, docs/configuration.md,
docs/lua-architecture.md, in-source comments in lua/{vm,mod,
runtimepath,init_lua,api/mod}.rs, satellite plugin's customization
comment, CLI long_about.
