+++ ttymap-tui/runtime/lua/plugin/aircraft/display.lua
