+++ ttymap-tui/runtime/lua/plugin/aircraft/opensky.lua
