+++ ttymap-tui/runtime/lua/plugin/ping_simulation.lua
