+++ ttymap-tui/runtime/lua/plugin/travel/routes/init.lua
@@ -5,7 +5,7 @@
 --
 --   1. Drop a `<country>.lua` next to this file with the same shape
 --      as `japan.lua`.
---   2. Append a `(require "travel.routes.<country>")` line to the
+--   2. Append a `(require "plugin.travel.routes.<country>")` line to the
 --      list below — note the parentheses, they are load-bearing
 --      (Lua 5.4 `require` returns the module *and* loader info; the
 --      parens truncate to a single value so the list doesn't grow
@@ -15,6 +15,6 @@
 -- `readdir`, and the explicit list keeps load order deterministic.
 
 return {
-    (require "travel.routes.japan"),
-    (require "travel.routes.italy"),
+    (require "plugin.travel.routes.japan"),
+    (require "plugin.travel.routes.italy"),
 }
