+++ ttymap-tui/runtime/lua/plugin/travel/routes/japan.lua
