+++ ttymap-tui/src/main.rs
@@ -14,7 +14,7 @@ use ttymap_tui::input::thread::InputHandle;
         with ANSI 256-color in your terminal.\n\n\
         Inspired by and based on mapscii (https://github.com/rastapasta/mapscii).\n\n\
         Config file:    ~/.config/ttymap/init.lua\n\
-        User plugins:   ~/.config/ttymap/plugin/  (activate via `require \"<name>\"` in init.lua)\n\
+        User plugins:   ~/.config/ttymap/lua/plugin/  (activate via `require \"plugin.<name>\"` in init.lua)\n\
         Bundled runtime: ~/.local/share/ttymap/lua/\n\
         Log file:       ~/.local/state/ttymap/ttymap.log\n\
         Tile cache:     ~/.cache/ttymap/",
